新しいページリストタブコンポーネントとヒーローセクションを追加し、OpenGraph画像生成機能を実装
